[feat]: rofi/darkman, Update to 1.7.2
Details:
 * Fix rofi configs after update 1.7.2
 * Add dark mode switching for rofi

Resolves:
 * Rofi 1.7.2 support

Tests:
 * Tested on Arch Linux

+++ .config/darkman/dark-mode.d/set-gtk-theme.sh
@@ -2,5 +2,11 @@
 
 xdg_config_path="$HOME/.config/xsettingsd/xsettingsd.conf"
 sed -i -e "s#-light#-dark#g" $xdg_config_path
-killall -HUP xsettingsd
+
+if ! pgrep -x "xsettingsd" > /dev/null
+then
+    xsettingsd &
+else
+    killall -HUP xsettingsd
+fi
